+++ src/FreeElephants/RestDaemon/Endpoint/Exception/InvalidCongurationValueException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FreeElephants\RestDaemon\Endpoint\Exception;
+
+use FreeElephants\RestDaemon\Exception\InvalidArgumentException;
+
+class InvalidCongurationValueException extends InvalidArgumentException
+{
+
+}