+++ tests/rest/CORSHeadersCept.php
@@ -1,0 +1,5 @@
+<?php
+$I = new RestTester($scenario);
+$I->wantToTest('CORS Headers. ');
+$I->sendHEAD('/');
+$I->seeHttpHeader('Access-Control-Allow-Origin', '*');